// src/libs/api/models/CartDetailResonse.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { Product } from './Product';

export type CartDetailResonse = {
    product_list?: Array<Product>;
    product_amount?: number;
    cost?: number;
    tax?: number;
    tax_ratio?: number;
    total_amount?: number;
};